Add message handling for mint v1
