+++ pkg/booking/booking.go
@@ -7,7 +7,6 @@ import (
 	"math/big"
 	"strings"
 
-	typesv2 "buf.build/gen/go/chain4travel/camino-messenger-protocol/protocolbuffers/go/cmp/types/v2"
 	"github.com/chain4travel/camino-messenger-contracts/go/contracts/cmaccount"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
@@ -129,16 +128,16 @@ func (bs *Service) BuyBookingToken(
 }
 
 // convertPriceToBigInt converts the price to its integer representation
-func (bs *Service) ConvertPriceToBigInt(price *typesv2.Price, totalDecimals int32) (*big.Int, error) {
+func (bs *Service) ConvertPriceToBigInt(value string, decimals int32, totalDecimals int32) (*big.Int, error) {
 	// Convert the value string to a big.Int
 	valueBigInt := new(big.Int)
-	_, ok := valueBigInt.SetString(price.Value, 10)
+	_, ok := valueBigInt.SetString(value, 10)
 	if !ok {
-		return nil, fmt.Errorf("failed to convert value to big.Int: %s", price.Value)
+		return nil, fmt.Errorf("failed to convert value to big.Int: %s", value)
 	}
 
 	// Calculate the multiplier as 10^(totalDecimals - price.Decimals)
-	multiplier := new(big.Int).Exp(big.NewInt(10), big.NewInt(int64(totalDecimals-price.Decimals)), nil)
+	multiplier := new(big.Int).Exp(big.NewInt(10), big.NewInt(int64(totalDecimals-decimals)), nil)
 
 	// Multiply the value by the multiplier
 	result := new(big.Int).Mul(valueBigInt, multiplier)
